Create very basic Unit Test structure.
Harder than I thought it was going to be. Edit mode tests have to be in a very specific location, and the asmdef file was needed as well to get the test runner to do anything. I don't understand fully why, but it works now and you can run tests with VS, Rider, or Unity (GUI or command line)